finished custom messages

+++ distance_service/src/client.cpp
@@ -14,12 +14,12 @@ int main(int argc, char **argv)
   ros::NodeHandle n;
   ros::ServiceClient client = n.serviceClient<distance_service::ComputeDistance>("compute_distance");
   distance_service::ComputeDistance srv;
-  srv.request.carX = atoll(argv[1]);
-  srv.request.carY = atoll(argv[2]);
-  srv.request.carZ = atoll(argv[3]);
-  srv.request.obsX = atoll(argv[4]);
-  srv.request.obsY = atoll(argv[5]);
-  srv.request.obsZ = atoll(argv[6]);
+  srv.request.carX = atof(argv[1]);
+  srv.request.carY = atof(argv[2]);
+  srv.request.carZ = atof(argv[3]);
+  srv.request.obsX = atof(argv[4]);
+  srv.request.obsY = atof(argv[5]);
+  srv.request.obsZ = atof(argv[6]);
   if (client.call(srv))
   {
     ROS_INFO("distance: %f", (double)srv.response.distance);
